Refactor: Combine MobileSideBar and SideBar into a single responsive component
- Created a new unified Sidebar.tsx component.
- Combined the JSX and logic from the previous mobile-sidebar.tsx and sidebar.tsx.
- Used Tailwind CSS (md:hidden, hidden md:flex) for responsive rendering of mobile and desktop views.
- Consolidated user data fetching using a single useGetUser() call.
- Ensured logout functionality, avatar display, and sidebar route rendering remain unchanged.
- Updated imports in Header.tsx and DashboardLayout.tsx to use the new Sidebar component.
- Deleted the old mobile-sidebar.tsx and sidebar.tsx files.
